added filtering and encoding

+++ src/main/java/zpi/service/AttractionServiceImpl.java
@@ -6,9 +6,10 @@ import org.springframework.stereotype.Service;
 import zpi.entity.Attraction;
 import zpi.repository.AttractionRepository;
 
+import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
-import java.util.Optional;
-
+import java.util.Set;
 
 @Service
 @RequiredArgsConstructor
@@ -28,33 +29,50 @@ public class AttractionServiceImpl implements AttractionService {
 
     @Override
     public Attraction updateAttraction(Attraction attraction, Integer attractionId) {
-        Attraction attractionDB
-                = attractionRepository.findById(attractionId)
-                .orElseThrow(() -> new EntityNotFoundException("Attraction not found for ID: " + attractionId));
-
-        if(attraction.getTitle() != null && !attraction.getTitle().isEmpty()) {
-            attractionDB.setTitle(attraction.getTitle());
+        if (!attractionRepository.existsById(attractionId)) {
+            throw new EntityNotFoundException("Attraction not found for ID: " + attractionId);
         }
-
-        if(attraction.getAttractionType() != null) {
-            attractionDB.setAttractionType(attraction.getAttractionType());
-        }
-
-        if(attraction.getDescription() != null && !attraction.getDescription().isEmpty()) {
-            attractionDB.setDescription(attraction.getDescription());
-        }
-
-        return attractionRepository.save(attractionDB);
+        attraction.setId(attractionId);
+        return attractionRepository.save(attraction);
     }
 
     @Override
     public void deleteAttractionById(Integer attractionId) {
+        if (!attractionRepository.existsById(attractionId)) {
+            throw new EntityNotFoundException("Attraction not found for ID: " + attractionId);
+        }
         attractionRepository.deleteById(attractionId);
     }
 
     @Override
     public Attraction findAttractionById(Integer attractionId) {
-        return attractionRepository.findById(attractionId).get();
+        return attractionRepository.findById(attractionId)
+                .orElseThrow(() -> new EntityNotFoundException("Attraction not found for ID: " + attractionId));
     }
+
+    @Override
+    public List<Attraction> getAttractionList(List<String> cities, List<String> districts, List<String> types) {
+        Set<Attraction> attractionsSet = new HashSet<>();
+
+        if (cities != null && !cities.isEmpty()) {
+            for (String city : cities) {
+                attractionsSet.addAll(attractionRepository.findByCityName(city));
+            }
+        }
+
+        if (districts != null && !districts.isEmpty()) {
+            for (String district : districts) {
+                attractionsSet.addAll(attractionRepository.findByDistrictName(district));
+            }
+        }
+
+        if (types != null && !types.isEmpty()) {
+            for (String type : types) {
+                attractionsSet.addAll(attractionRepository.findByType(type));
+            }
+        }
+        return new ArrayList<>(attractionsSet);
+    }
+
 
 }
